Renames `generate/testgen` package to `generate/tests`
