+++ internal/generate/tests/common.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package testgen
+package tests
 
 import (
 	_ "embed"
